Create awx/main/utils/redis.py and move Redis client functions to avoid circular imports
