+++ awx/main/tests/functional/conftest.py
@@ -117,13 +117,9 @@ class FakeRedis:
 
 
 @pytest.fixture
-def fake_redis(request):
-    # Skip this fixture if the test is marked with donotusefakeredis
-    if 'donotusefakeredis' in request.keywords:
+def fake_redis():
+    with mock.patch('redis.Redis', new=FakeRedis):  # turn off redis stuff
         yield
-    else:
-        with mock.patch('redis.Redis', new=FakeRedis):  # turn off redis stuff
-            yield
 
 
 @pytest.fixture
