+++ pytest.ini
@@ -11,7 +11,6 @@ markers =
     activity_stream_access:
     job_runtime_vars:
     fixture_args:
-    donotusefakeredis: skip FakeRedis fixture for this test
 
 filterwarnings =
   error
